Restructure side menu into mobile menu and side menu, move to subfolder
