+++ application/account-management/WebApp/shared/components/sideMenu/NavigationMenuItems.tsx
@@ -1,0 +1,49 @@
+import { t } from "@lingui/core/macro";
+import { Trans } from "@lingui/react/macro";
+import { FederatedMenuButton, SideMenuSeparator } from "@repo/ui/components/SideMenu";
+import { BoxIcon, CircleUserIcon, HomeIcon, UsersIcon } from "lucide-react";
+import type { SharedSideMenuProps } from "./SharedSideMenu";
+
+// Navigation items shared between mobile and desktop menus
+export function NavigationMenuItems({
+  currentSystem
+}: Readonly<{ currentSystem: SharedSideMenuProps["currentSystem"] }>) {
+  return (
+    <>
+      <FederatedMenuButton
+        icon={HomeIcon}
+        label={t`Home`}
+        href="/admin"
+        isCurrentSystem={currentSystem === "account-management"}
+      />
+
+      <SideMenuSeparator>
+        <Trans>Organization</Trans>
+      </SideMenuSeparator>
+
+      <FederatedMenuButton
+        icon={CircleUserIcon}
+        label={t`Account`}
+        href="/admin/account"
+        isCurrentSystem={currentSystem === "account-management"}
+      />
+      <FederatedMenuButton
+        icon={UsersIcon}
+        label={t`Users`}
+        href="/admin/users"
+        isCurrentSystem={currentSystem === "account-management"}
+      />
+
+      <SideMenuSeparator>
+        <Trans>Back Office</Trans>
+      </SideMenuSeparator>
+
+      <FederatedMenuButton
+        icon={BoxIcon}
+        label={t`Dashboard`}
+        href="/back-office"
+        isCurrentSystem={currentSystem === "back-office"}
+      />
+    </>
+  );
+}
